Refactor main.ts to add support for adding a list of numbers
This commit refactors the main.ts file in the wasm-node project to include support for adding a list of numbers using the newly added `addList` function from the wasm-rust library. It also adds an example usage of the `addList` function by adding an array of numbers and logging the result.